Correct decoding strategy
Instead of decoding until end-of-sentence token ";", we decode until the new aux point is defined by two predicates, and while decoding this sentence, we mask out the sentence ending tokens (";" and "."). At the end, then, we have to append eos " ;" manually.

with this change, we matched the decoding strategy of the internal codebase and reproduced the result.